backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.18
passlib==1.7.4
bcrypt==3.2.2
PyJWT==2.8.0
python-multipart==0.0.9
pydantic==2.5.0
email-validator==2.1.0
